+++ src/main/java/cn/nukkit/item/ItemBoat.java
@@ -47,7 +47,7 @@ public class ItemBoat extends Item {
     }
 
     @Override
-    public boolean onActivate(Level level, Player player, Block block, Block target, BlockFace face, double fx, double fy, double fz) {
+    public boolean onActivate(Level level, Player player, Block block, Block target, BlockFace face, float fx, float fy, float fz) {
         if (face != BlockFace.UP || block.isLiquid() || !block.isAir() && block.canContainWater() && level.getExtraBlock(block).isWater()) {
             return false;
         }
