+++ src/main/java/cn/nukkit/item/ItemFireCharge.java
@@ -30,7 +30,7 @@ public class ItemFireCharge extends Item {
     }
 
     @Override
-    public boolean onActivate(Level level, Player player, Block block, Block target, BlockFace face, double fx, double fy, double fz) {
+    public boolean onActivate(Level level, Player player, Block block, Block target, BlockFace face, float fx, float fy, float fz) {
 //        if (player.isAdventure()) {
 //            return false;
 //        }
