+++ src/main/java/cn/nukkit/item/ItemMinecartCommandBlock.java
@@ -25,7 +25,7 @@ public class ItemMinecartCommandBlock extends Item {
     }
 
     @Override
-    public boolean onActivate(Level level, Player player, Block block, Block target, BlockFace face, double fx, double fy, double fz) {
+    public boolean onActivate(Level level, Player player, Block block, Block target, BlockFace face, float fx, float fy, float fz) {
         //TODO
         return false;
     }
